Fix Int overflow and settingVC bug
In release mode, Int is 32 bit, so a phone number will be overflow.
Change Int to UInt64

In settingVC, we deleted one session so the didSelect need to change

+++ swift-demo/Hyphenate Messenger/LoginPhoneNumberViewController.swift
@@ -90,7 +90,7 @@ class LoginPhoneNumberViewController: UIViewController, UITextFieldDelegate {
         let verifyVC = segue.destination as! LoginVerificationViewController
         // setting the phone number and operation mode of verify VC
         // mode will determine whether VC calls signup or login for hyphenate
-        verifyVC.phoneNumber = Int((numberField?.text)!)!
+        verifyVC.phoneNumber = UInt64(Int((numberField?.text)!)!)
         verifyVC.mode = mode
     }
